requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
Async-OAuthlib==0.0.9
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloud-sql-python-connector==1.0.0
colorama==0.4.6
cryptography==38.0.4
discord==2.1.0
discord.py==2.1.0
Flask==2.2.2
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud==0.34.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.21.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.3
oauthlib==3.2.2
priority==2.0.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
quart==0.18.3
Quart-Discord==2.1.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.45
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.2
